로그인 수정

+++ StageLink-1/src/main/java/com/pro/stagelink/repository/MemberRepository.java
@@ -4,10 +4,10 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.pro.stagelink.domain.Member;
+import com.pro.stagelink.entity.Member;
 import com.pro.stagelink.dto.MemberDTO;
 
-public interface MemberRepository extends JpaRepository<Member, Integer> {
+public interface MemberRepository extends JpaRepository<Member, Long> {
 
     // ✅ 이름 기준 부분 검색 + 페이징
     Page<Member> findByNameContaining(String name, Pageable pageable);
